rename yes module to yes_no to work around mypy issue
