+++ src/borg/helpers/__init__.py
@@ -16,7 +16,7 @@ from .parseformat import *  # NOQA
 from .process import *  # NOQA
 from .progress import *  # NOQA
 from .time import *  # NOQA
-from .yes import *  # NOQA
+from .yes_no import *  # NOQA
 
 from .msgpack import is_slow_msgpack, is_supported_msgpack, int_to_bigint, bigint_to_int, get_limited_unpacker
 from . import msgpack
